sensors/requirements.txt
requests
websocket-client
grpcio
grpcio-tools